modal user에 따라 반응
