+++ src/components/common/Product.jsx
@@ -4,6 +4,7 @@ import { useSelector } from 'react-redux';
 import { useState } from 'react';
 import { MyProductModal } from '../common/Modal';
 
+
 const ProductItemBox = styled.div`
     width: 140px;
     margin-right: 10px;
@@ -40,38 +41,29 @@ const ProductItemBox = styled.div`
     }
 `;
 
-const Product = ({ name, price, src, id }) => {
+const Product = ({ name, price, src, itemLink, writerId }) => {
     const userId = useSelector(state => state.auth.id);
-    // // 모달
-    // const [modalOn, setModalOn] = useState(false);
-    // function openModal() {
-    //     setModalOn(true);
-    // }
-    // function closeModal() {
-    //     setModalOn(false);
-    // }
+    const [modalOn, setModalOn] = useState(false);
 
-    const handler = () => {
-        console.log('상품등록유저', id);
+    function openModal() {
+        console.log('상품등록유저', writerId);
         console.log('유저:', userId);
+        if (userId !== writerId) {
+            setModalOn(false);
+            console.log('상품링크로 이동')
+            window.open({ itemLink }, '_blank')
+        } else {
+            setModalOn(true);
+        }
+    }
 
-        // if (userId === id) {
-        //     console.log('같다')
-        //     return (
-        //         <>
-        //             {modalOn === true ? <MyProductModal openModal={openModal} closeModal={closeModal} /> : ''}
-        //         </>
-        //     )
-        // } else {
-        //     return (
-        //         console.log('링크로 이동하기')
-        //     )
-        // }
+    function closeModal() {
+        setModalOn(false);
     }
 
     return (
         <>
-            <ProductItemBox onClick={handler}>
+            <ProductItemBox onClick={openModal}>
                 <img
                     className='img-product'
                     src={src}
@@ -80,6 +72,7 @@ const Product = ({ name, price, src, id }) => {
                 <p className='txt-productName'>{name}</p>
                 <span className='txt-productPrice'>{price}원</span>
             </ProductItemBox>
+            {modalOn === true ? <MyProductModal openModal={openModal} closeModal={closeModal} /> : ''}
         </>
     );
 };
